🚧 wip: add course dialog with tee form
Moved add tee dialog content to the content of the second page in the add course multi page dialog. Need to make it into a form for easier validation before submission. Also need to abstract the add tee content to a separate component to be reused for adding a tee to an existing course.